training/requirements.txt
torch>=2.0.0
python-chess>=1.9.4

